feature-17: Changed Next.js app folders structure by moving everything to the app folder
