+++ client/app/(routes)/favicon.ico
