+++ client/app/(routes)/layout.js
@@ -2,7 +2,7 @@ import { Geist, Geist_Mono } from 'next/font/google'
 import { ToastContainer } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
 import './globals.css'
-import Navbar from '@/components/Navbar'
+import Navbar from '@components/Navbar'
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
