+++ client/app/(routes)/page.js
@@ -2,15 +2,17 @@
 
 import { useEffect, useState } from 'react'
 import { toast } from 'react-toastify'
-import { API_URL } from '@/config/config'
-import Quotes from '@/components/Quotes'
-import Button from '@/components/Button'
+import { ClipLoader } from 'react-spinners'
+import { API_URL } from '@config/config'
+import Quotes from '@components/Quotes'
+import Button from '@components/Button'
 
 const RANDOM_QUOTES_LIMIT = 9
 const RANDOM_QUOTES_URL = `${API_URL}/quotes/random?limit=${RANDOM_QUOTES_LIMIT}`
 
 export default function RandomQuotesPage() {
   const [quotes, setQuotes] = useState([])
+  const [isLoading, setIsLoading] = useState(true)
 
   const fetchQuotes = async () => {
     try {
@@ -20,6 +22,8 @@ export default function RandomQuotesPage() {
     } catch (error) {
       toast.error(error.message)
       console.log('Error fetching quotes: ', error)
+    } finally {
+      setIsLoading(false)
     }
   }
 
@@ -29,8 +33,14 @@ export default function RandomQuotesPage() {
 
   return (
     <div className="p-4">
-      <Button onClick={fetchQuotes} text='Get random quotes' />
-      <Quotes quotes={quotes} />
+      <Button onClick={fetchQuotes} text="Get random quotes" />
+      {isLoading ? (
+        <div className="flex justify-center items-center pt-20">
+          <ClipLoader size={60} color="violet" />
+        </div>
+      ) : (
+        <Quotes quotes={quotes} />
+      )}
     </div>
   )
 }
