+++ client/app/(routes)/quotes/[id]/edit/page.js
@@ -4,10 +4,10 @@ import { useEffect, useState, use } from 'react'
 import { useRouter } from 'next/navigation'
 import { toast } from 'react-toastify'
 import { ClipLoader } from 'react-spinners'
-import { API_URL } from '@/config/config'
-import InputField from '@/components/InputField'
-import Button from '@/components/Button'
-import { isFormValid } from '../../utils/validation'
+import { API_URL } from '@config/config'
+import { isFormValid } from '@utils/validation'
+import InputField from '@components/InputField'
+import Button from '@components/Button'
 
 export default function EditQuotePage({ params }) {
   const { id } = use(params)
