+++ client/app/(routes)/search/page.js
@@ -4,10 +4,10 @@ import { useEffect, useState } from 'react'
 import { useRouter, useSearchParams } from 'next/navigation'
 import { ClipLoader } from 'react-spinners'
 import { toast } from 'react-toastify'
-import { API_URL } from '@/config/config'
-import Quotes from '@/components/Quotes'
-import InputField from '@/components/InputField'
-import Button from '@/components/Button'
+import { API_URL } from '@config/config'
+import Quotes from '@components/Quotes'
+import InputField from '@components/InputField'
+import Button from '@components/Button'
 
 const QUOTES_URL = `${API_URL}/quotes`
 const CATEGORY_NAME_REGEX = /^[a-z0-9\-]+$/
@@ -194,8 +194,8 @@ export default function SearchQuotesPage() {
       </div>
 
       <div className="flex justify-center md-6 space-x-4">
-        <Button onClick={handleSearch} text='Search' />
-        <Button onClick={clearSearch} text='Clear' variant="secondary" />
+        <Button onClick={handleSearch} text="Search" />
+        <Button onClick={clearSearch} text="Clear" variant="secondary" />
       </div>
 
       {isLoading ? (
