+++ client/app/_components/Quote.jsx
