+++ client/app/_components/Quotes.jsx
